+++ .hintrc
@@ -12,7 +12,6 @@
     "disown-opener",
     "html-checker",
     "meta-charset-utf-8",
-    "meta-viewport",
-    "no-inline-styles:error"
+    "meta-viewport"
   ]
 }
