+++ src/index.html
@@ -32,8 +32,8 @@
       </section>
     </main>
     <footer>
-      <p>Created by <a href="https://github.com/GoddessNgozi" target="_blank">Goddess Ngozi </a> and 
-        <a href="https://github.com/abe1able" target="_blank">Abel Seifu</a></p>
+      <p>Created by <a href="https://github.com/GoddessNgozi">Goddess Ngozi </a> and 
+        <a href="https://github.com/abe1able">Abel Seifu</a></p>
     </footer>
 </body>
 </html>
