+++ src/style.css
@@ -1,171 +1,174 @@
 @import url('https://fonts.googleapis.com/css2?family=Nunito:ital,wght@0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap');
 
 * {
-    box-sizing: border-box;
-    margin: 0;
-    list-style: none;
+  box-sizing: border-box;
+  margin: 0;
+  list-style: none;
 }
 
 body {
-    background-image: url('./images/cinema5.jpg');
-    /* background-image: linear-gradient(rgba(0, 0, 0, 0.8), rgba(0, 0, 0, 0.2)), url('./images/cinema5.jpg'); */
-    background-size: cover; 
-    background-repeat: no-repeat;
-    background-attachment: fixed;
-    color: white;
-    font-family: 'Nunito', sans-serif;
+  background-image: url('./images/cinema5.jpg');
+
+  /* background-image: linear-gradient(rgba(0, 0, 0, 0.8), rgba(0, 0, 0, 0.2)), url('./images/cinema5.jpg'); */
+  background-size: cover;
+  background-repeat: no-repeat;
+  background-attachment: fixed;
+  color: white;
+  font-family: 'Nunito', sans-serif;
 }
 
 header {
-    background-color: rgba(215, 243, 31, 0);
-    /* backdrop-filter: saturate(180%) blur(10px); */
-    backdrop-filter: blur(10px);
-    position: fixed;
-    top: 0;
-    width: 100%;
-    z-index: 20;
+  background-color: rgba(215, 243, 31, 0);
+
+  /* backdrop-filter: saturate(180%) blur(10px); */
+  backdrop-filter: blur(10px);
+  position: fixed;
+  top: 0;
+  width: 100%;
+  z-index: 20;
 }
 
 nav {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
 }
 
 #searchInput {
-    height: 3rem;
-    width: 100%;
-    background-color: transparent;
-    border: none;
-    color: white;
-    outline: none;
+  height: 3rem;
+  width: 100%;
+  background-color: transparent;
+  border: none;
+  color: white;
+  outline: none;
 }
 
 #searchIcon {
-    font-size: 2.025rem;
-    background-color: rgba(214,243,31,.9);
-    padding: 2px;
-    border-radius: 2rem;
+  font-size: 2.025rem;
+  background-color: rgba(214, 243, 31, 0.9);
+  padding: 2px;
+  border-radius: 2rem;
 }
 
 .search-bar {
-    display: flex;
-    border: 1px solid rgba(214,243,31,.9);
-    padding: 5px;
-    border-radius: 1rem;
-    width: 30rem;
-    background-color: rgba(215, 243, 31, 0.287);
+  display: flex;
+  border: 1px solid rgba(214, 243, 31, 0.9);
+  padding: 5px;
+  border-radius: 1rem;
+  width: 30rem;
+  background-color: rgba(215, 243, 31, 0.287);
 }
 
 .logo {
-    font-size: 2.51rem;
-    color: rgba(214,243,31,.9);
+  font-size: 2.51rem;
+  color: rgba(214, 243, 31, 0.9);
 }
 
 .right-nav {
-    display: flex;
-    align-items: center;
-    gap: 10px;
+  display: flex;
+  align-items: center;
+  gap: 10px;
 }
 
 .shows-counter {
-    /*border: 1px solid rgba(214,243,31,.9);
+  /* border: 1px solid rgba(214,243,31,.9);
     padding: 5px;
     border-radius: 1rem;
     background-color: rgba(215, 243, 31, 0.287);
-    color: rgba(214,243,31,.9);*/
-    padding: 10px 7px 10px 7px;
-    border-radius: 1rem;
-    background-color: rgba(214,243,31,.9);
-    color: black;
+    color: rgba(214,243,31,.9); */
+  padding: 10px 7px 10px 7px;
+  border-radius: 1rem;
+  background-color: rgba(214, 243, 31, 0.9);
+  color: black;
 }
 
 .card {
-    display: flex;
-    justify-content: space-around;
-    align-items: flex-end;
-    background-size: 100% 100%;
-    background-repeat: no-repeat;
-    height: 18rem;
-    width: 13rem;
-    border-radius: 1.5rem;
-    border: 1px solid rgba(214,243,31,.9);
+  display: flex;
+  justify-content: space-around;
+  align-items: flex-end;
+  background-size: 100% 100%;
+  background-repeat: no-repeat;
+  height: 18rem;
+  width: 13rem;
+  border-radius: 1.5rem;
+  border: 1px solid rgba(214, 243, 31, 0.9);
 }
 
 .show-card {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin-block-end: 30px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-block-end: 30px;
 }
 
 .show-cards {
-    display: flex;
-    justify-content: flex-end;
-    margin-top: 150px;
+  display: flex;
+  justify-content: flex-end;
+  margin-top: 150px;
 }
 
 .cards {
-    display: flex;
-    justify-content: center;
-    flex-wrap: wrap;
-    gap: 1.25rem;
-    width: 63.75%;
-    padding-right: 3.75%;
+  display: flex;
+  justify-content: center;
+  flex-wrap: wrap;
+  gap: 1.25rem;
+  width: 63.75%;
+  padding-right: 3.75%;
 }
 
-.rating, .likes, .comment-btn {
-    margin-bottom: 10px;
-    color: black;
-    border-radius: 9999px;
-    border: 2px solid rgba(214,243,31,.9);
-    background-color: rgba(255, 255, 255, 0.696);
-    padding: 0.25rem;
+.rating,
+.likes,
+.comment-btn {
+  margin-bottom: 10px;
+  color: black;
+  border-radius: 9999px;
+  border: 2px solid rgba(214, 243, 31, 0.9);
+  background-color: rgba(255, 255, 255, 0.696);
+  padding: 0.25rem;
 }
 
 .status {
-    color: black;
-    border-radius: 9999px;
-    border: 4px solid rgba(1, 1, 0, 0.416);
-    background-color: rgb(226, 250, 89);
-    max-width: 10rem;
-    padding: 0.5rem;
-    font-size: 1.125rem;
-    line-height: 1.75rem;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    z-index: 5;
-    margin-bottom: -25px;
+  color: black;
+  border-radius: 9999px;
+  border: 4px solid rgba(1, 1, 0, 0.416);
+  background-color: rgb(226, 250, 89);
+  max-width: 10rem;
+  padding: 0.5rem;
+  font-size: 1.125rem;
+  line-height: 1.75rem;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  z-index: 5;
+  margin-bottom: -25px;
 }
 
 .comment-btn {
-    font-size: 1rem;
+  font-size: 1rem;
 }
 
 .show-name {
-    color: white;
-    font-size: large;
-    text-align: center;
-    max-width: 13rem;
+  color: white;
+  font-size: large;
+  text-align: center;
+  max-width: 13rem;
 }
 
 .likes {
-    cursor: pointer;
+  cursor: pointer;
 }
 
 .comments {
-    display: flex;
-    justify-content: center;
-    margin-top: 5px;
+  display: flex;
+  justify-content: center;
+  margin-top: 5px;
 }
 
 footer {
   display: flex;
   align-items: flex-end;
   justify-content: center;
-  background-color: rgba(215, 243, 31, 0);
   width: 100%;
   height: 2.5rem;
   margin-top: 50px;
@@ -173,21 +176,21 @@ footer {
   background-color: rgba(215, 243, 31, 0);
 }
 
-footer p { 
-    font-size: 1.225rem;
-    color: white;
+footer p {
+  font-size: 1.225rem;
+  color: white;
 }
 
 footer p a {
-    color: white;
-    text-decoration: none;
-    cursor: pointer;
+  color: white;
+  text-decoration: none;
+  cursor: pointer;
 }
 
 @media (min-width: 768px) {
-    nav {
-        padding-left: 6%;
-        padding-right: 6%;
-        padding-top: 0.5%;
-    }
+  nav {
+    padding-left: 6%;
+    padding-right: 6%;
+    padding-top: 0.5%;
+  }
 }
