+++ webpack.config.js
@@ -2,7 +2,7 @@ const path = require('path');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-  mode: 'development',  
+  mode: 'development',
   entry: './src/index.js',
   devServer: {
     static: './dist',
